requirements.txt
numpy==1.21.2
pandas==1.3.2
tqdm==4.62.2
torch==1.9.0
PyYAML==5.4.1
pynvml==11.4.1
sklearn==0.0
matplotlib==3.4.3
